Adapt remaining test cases

+++ projects/hipcub/test/hipcub/test_hipcub_block_adjacent_difference.cpp
@@ -263,7 +263,7 @@ TYPED_TEST(HipcubBlockAdjacentDifferenceSubtract, SubtractLeft)
     static constexpr int size = items_per_block * 20;
     static constexpr int grid_size = size / items_per_block;
 
-	CHECK_SIZE_ENABLEMENT(size);
+	CHECK_SIZE_ENABLEMENT(static_cast<size_t>(size));
 
     // Given block size not supported
     if(block_size > test_utils::get_max_block_size())
@@ -367,7 +367,7 @@ TYPED_TEST(HipcubBlockAdjacentDifferenceSubtract, SubtractLeftPartialTile)
     static constexpr int size = items_per_block * 20;
     static constexpr int grid_size = size / items_per_block;
 
-	CHECK_SIZE_ENABLEMENT(size);
+	CHECK_SIZE_ENABLEMENT(static_cast<size_t>(size));
 
     // Given block size not supported
     if(block_size > test_utils::get_max_block_size())
@@ -491,7 +491,7 @@ TYPED_TEST(HipcubBlockAdjacentDifferenceSubtract, SubtractRight)
     static constexpr int size = items_per_block * 20;
     static constexpr int grid_size = size / items_per_block;
 
-	CHECK_SIZE_ENABLEMENT(size);
+	CHECK_SIZE_ENABLEMENT(static_cast<size_t>(size));
 
     // Given block size not supported
     if(block_size > test_utils::get_max_block_size())
@@ -595,7 +595,7 @@ TYPED_TEST(HipcubBlockAdjacentDifferenceSubtract, SubtractRightPartialTile)
     static constexpr int size = items_per_block * 20;
     static constexpr int grid_size = size / items_per_block;
 
-	CHECK_SIZE_ENABLEMENT(size);
+	CHECK_SIZE_ENABLEMENT(static_cast<size_t>(size));
 
     // Given block size not supported
     if(block_size > test_utils::get_max_block_size())
